# packages/api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Public REST API of Honest Robin: Time. Authenticate with a personal access token (`Authorization: Bearer hrt_…`). Session-authenticated browser clients send the `HonestRobin-Account-Id` header to pick an account.",
    "license" : {
      "name" : "AGPL-3.0-only",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "title" : "Honest Robin: Time API",
    "version" : "v1"
  },
  "security" : [ {
    "token" : [ ]
  } ],
...
  "paths" : {
    "/api/v1/account" : {
      "get" : {
        "operationId" : "account_get",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "account" ]
      },
      "patch" : {
        "operationId" : "account_update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountUpdate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "account" ]
      }
    },
...
    "/api/v1/accounts" : {
      "post" : {
        "operationId" : "account_create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewAccount"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedAccount"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "account" ]
      }
    },
...
    "/api/v1/me" : {
      "get" : {
        "operationId" : "me_get",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me" ]
      },
      "patch" : {
        "operationId" : "me_update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeUpdate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me" ]
      }
    },
    "/api/v1/me/api_tokens" : {
      "get" : {
        "operationId" : "me_api_tokens",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ApiTokenView"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me" ]
      },
      "post" : {
        "operationId" : "me_create_api_token",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewApiToken"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedApiToken"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "me" ]
      }
    },
    "/api/v1/me/api_tokens/{id}" : {
      "delete" : {
        "operationId" : "me_revoke_api_token",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "me" ]
      }
    },
...
    "/api/v1/me/password" : {
      "post" : {
        "operationId" : "me_change_password",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordChange"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "me" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AccountUpdate" : {
        "properties" : {
          "address_line1" : {
            "type" : "string"
          },
          "address_line2" : {
            "type" : "string"
          },
          "approvals_enabled" : {
            "type" : "boolean"
          },
          "bic" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "company_reg_no" : {
            "type" : "string"
          },
          "country_code" : {
            "type" : "string"
          },
          "default_currency" : {
            "type" : "string"
          },
          "duration_format" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "legal_name" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "peppol_id" : {
            "type" : "string"
          },
          "peppol_scheme" : {
            "type" : "string"
          },
          "postal_code" : {
            "type" : "string"
          },
          "region" : {
            "type" : "string"
          },
          "require_two_factor" : {
            "type" : "boolean"
          },
          "time_rounding_minutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "time_rounding_mode" : {
            "type" : "string"
          },
          "timesheet_reminders_enabled" : {
            "type" : "boolean"
          },
          "timezone" : {
            "type" : "string"
          },
          "vat_id" : {
            "type" : "string"
          },
          "week_start" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "AccountView" : {
        "properties" : {
          "address_line1" : {
            "type" : "string"
          },
          "address_line2" : {
            "type" : "string"
          },
          "approvals_enabled" : {
            "type" : "boolean"
          },
          "bic" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "company_reg_no" : {
            "type" : "string"
          },
          "country_code" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "default_currency" : {
            "type" : "string"
          },
          "deletes_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "duration_format" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "legal_name" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "peppol_id" : {
            "type" : "string"
          },
          "peppol_scheme" : {
            "type" : "string"
          },
          "postal_code" : {
            "type" : "string"
          },
          "region" : {
            "type" : "string"
          },
          "require_two_factor" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          },
          "time_rounding_minutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "time_rounding_mode" : {
            "type" : "string"
          },
          "timesheet_reminders_enabled" : {
            "type" : "boolean"
          },
          "timezone" : {
            "type" : "string"
          },
          "updated_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "vat_id" : {
            "type" : "string"
          },
          "week_start" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "approvals_enabled", "created_at", "default_currency", "duration_format", "id", "locale", "name", "require_two_factor", "status", "time_rounding_minutes", "time_rounding_mode", "timesheet_reminders_enabled", "timezone", "updated_at", "week_start" ]
      },
...
      "ApiTokenView" : {
        "properties" : {
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expires_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "last_used_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "name" : {
            "type" : "string"
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "token_hint" : {
            "type" : "string"
          }
        },
        "required" : [ "created_at", "id", "name", "scopes", "token_hint" ]
      },
...
      "CreatedAccount" : {
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "id" ]
      },
      "CreatedApiToken" : {
        "properties" : {
          "api_token" : {
            "$ref" : "#/components/schemas/ApiTokenView"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "api_token", "token" ]
      },
...
      "MeUpdate" : {
        "properties" : {
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "MeView" : {
        "properties" : {
          "accounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MyAccount"
            }
          },
          "current_account_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "current_membership_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "email" : {
            "type" : "string"
          },
          "email_verification_required" : {
            "type" : "boolean"
          },
          "email_verified" : {
            "type" : "boolean"
          },
          "has_password" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "is_instance_admin" : {
            "type" : "boolean"
          },
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/Permissions"
            } ]
          },
          "role" : {
            "type" : "string"
          },
          "two_factor_enabled" : {
            "type" : "boolean"
          },
          "two_factor_setup_required" : {
            "type" : "boolean"
          }
        },
        "required" : [ "accounts", "email", "email_verification_required", "email_verified", "has_password", "id", "is_instance_admin", "locale", "name", "two_factor_enabled", "two_factor_setup_required" ]
      },
      "MyAccount" : {
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "membership_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "membership_id", "name", "role", "status" ]
      },
...
      "NewAccount" : {
        "properties" : {
          "default_currency" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          },
          "week_start" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "name" ]
      },
      "NewApiToken" : {
        "properties" : {
          "expires_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "name" : {
            "type" : "string"
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "name", "scopes" ]
      },
...
      "PasswordChange" : {
        "properties" : {
          "current_password" : {
            "type" : "string"
          },
          "new_password" : {
            "type" : "string"
          }
        },
        "required" : [ "new_password" ]
      },
...
      "Permissions" : {
        "properties" : {
          "can_manage_invoices" : {
            "type" : "boolean"
          },
          "can_manage_projects" : {
            "type" : "boolean"
          },
          "can_see_rates" : {
            "type" : "boolean"
          }
        },
        "required" : [ "can_manage_invoices", "can_manage_projects", "can_see_rates" ]
      },
...
    },
    "securitySchemes" : {
      "token" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}